refactor: reorganize folder structure (#53)
