+++ lib/constants/sports.dart
@@ -1,4 +1,4 @@
-import 'package:isis3510_team32_flutter/models/sport_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/sport_model.dart';
 
 Map<String, SportModel> initiationSports = {
   "basketball": SportModel(
