+++ lib/models/repositories/auth_repository.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:isis3510_team32_flutter/models/user_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/user_model.dart';
 
 class AuthRepository {
   static const String userCollection = "users";
