+++ lib/models/repositories/home_repository.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:isis3510_team32_flutter/models/booking_model.dart';
-import 'package:isis3510_team32_flutter/models/user_model.dart';
-import 'package:isis3510_team32_flutter/models/venue_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/booking_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/user_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/venue_model.dart';
 import 'package:isis3510_team32_flutter/core/firebase_service.dart';
 
 class HomeRepository {
