+++ lib/view_models/auth/auth_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:isis3510_team32_flutter/models/user_model.dart';
-import 'package:isis3510_team32_flutter/repositories/auth_repository.dart';
+import 'package:isis3510_team32_flutter/models/data_models/user_model.dart';
+import 'package:isis3510_team32_flutter/models/repositories/auth_repository.dart';
 import 'package:isis3510_team32_flutter/view_models/auth/auth_event.dart';
 import 'package:isis3510_team32_flutter/view_models/auth/auth_state.dart';
