+++ lib/view_models/auth/auth_event.dart
@@ -1,5 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:isis3510_team32_flutter/models/user_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/user_model.dart';
 
 abstract class AuthEvent {}
