+++ lib/view_models/home/home_bloc.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:isis3510_team32_flutter/models/booking_model.dart';
-import 'package:isis3510_team32_flutter/models/user_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/booking_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/user_model.dart';
 import 'package:isis3510_team32_flutter/view_models/auth/auth_bloc.dart';
 import 'package:isis3510_team32_flutter/view_models/auth/auth_state.dart';
-import 'package:isis3510_team32_flutter/repositories/home_repository.dart';
+import 'package:isis3510_team32_flutter/models/repositories/home_repository.dart';
 
 part 'home_event.dart';
 part 'home_state.dart';
