+++ lib/view_models/initiation/initiation_state.dart
@@ -1,4 +1,4 @@
-import 'package:isis3510_team32_flutter/models/sport_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/sport_model.dart';
 
 class InitiationState {
   String? name;
