+++ lib/views/search_view.dart
@@ -5,7 +5,7 @@ import 'package:isis3510_team32_flutter/widgets/search_view_widgets/sport_button
 import 'package:isis3510_team32_flutter/core/app_colors.dart';
 import 'package:isis3510_team32_flutter/view_models/search/search_bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:isis3510_team32_flutter/models/sport_model.dart';
+import 'package:isis3510_team32_flutter/models/data_models/sport_model.dart';
 
 class SearchView extends StatelessWidget {
   const SearchView({super.key});
